Move Event Log panel under system
