--- web/src/panels/hri/EventLogPanel.jsx
@@ -1,9 +1,0 @@
-import EventLog from '../../panels/EventLog';
-import '../../tabs/HRITab.css';
-
-function EventLogPanel() {
-  return <EventLog />;
-}
-
-export default EventLogPanel;
-export { EventLogPanel };
