+++ web/src/panels/system/EventLogPanel.css
